Habitalia consulting

+++ src/main/java/com/nautik/api/domain/moorings/MooringDimension.java
@@ -19,11 +19,11 @@ public class MooringDimension {
 
     @NotNull
     @Column(name = "max_length", nullable = false)
-    private Long maxLength;
+    private Integer maxLength;
 
     @NotNull
     @Column(name = "max_beam", nullable = false)
-    private Long maxBeam;
+    private Integer maxBeam;
 
 
 }
